requirements.txt
pandas
numpy
pymupdf
python-docx
python-pptx
markdown
langchain
chromadb
google-generativeai
streamlit
